refactor: rename rpc into bus
