+++ src/farcasterd/trade_state_machine.rs
@@ -1,18 +1,18 @@
-use crate::farcasterd::runtime::{launch, launch_swapd, syncer_up, Runtime};
-use crate::rpc::ctl::Ctl;
-use crate::rpc::ctl::{ProtoPublicOffer, PubOffer};
-use crate::rpc::msg::{Msg, TakeCommit};
-use crate::rpc::request::{
+use crate::bus::ctl::Ctl;
+use crate::bus::ctl::{ProtoPublicOffer, PubOffer};
+use crate::bus::msg::{Msg, TakeCommit};
+use crate::bus::request::{
     BitcoinAddress, BitcoinFundingInfo, CheckpointEntry, FundingInfo, InitSwap, LaunchSwap,
     MadeOffer, MoneroAddress, MoneroFundingInfo, OfferInfo, OfferStatus, OfferStatusPair,
     TookOffer,
 };
-use crate::rpc::{Failure, FailureCode};
+use crate::bus::{Failure, FailureCode};
+use crate::farcasterd::runtime::{launch, launch_swapd, syncer_up, Runtime};
 use crate::LogStyle;
 use crate::{
+    bus::request::{Outcome, Request},
     error::Error,
     event::{Event, StateMachine},
-    rpc::request::{Outcome, Request},
     ServiceId,
 };
 use bitcoin::hashes::hex::ToHex;
