+++ src/lib.rs
@@ -45,6 +45,8 @@ extern crate serde_crate as serde;
 #[macro_use]
 extern crate serde_with;
 
+#[cfg(feature = "_rpc")]
+pub mod bus;
 #[cfg(feature = "cli")]
 pub mod cli;
 #[cfg(feature = "_rpc")]
@@ -53,8 +55,6 @@ pub mod error;
 pub mod event;
 #[cfg(feature = "shell")]
 pub mod opts;
-#[cfg(feature = "_rpc")]
-pub mod rpc;
 
 #[cfg(feature = "node")]
 pub mod databased;
