+++ src/swapd/swap_state.rs
@@ -1,8 +1,8 @@
 use farcaster_core::role::{SwapRole, TradeRole};
 use strict_encoding::{StrictDecode, StrictEncode};
 
-use crate::rpc::msg::Commit;
-use crate::rpc::request::{Outcome, Params};
+use crate::bus::msg::Commit;
+use crate::bus::request::{Outcome, Params};
 
 #[derive(Display, Debug, Clone, StrictEncode, StrictDecode)]
 pub enum AliceState {
