+++ src/syncerd/bitcoin_syncer.rs
@@ -1,6 +1,6 @@
+use crate::bus::request::SyncerdBridgeEvent;
+use crate::bus::Request;
 use crate::error::SyncerError;
-use crate::rpc::request::SyncerdBridgeEvent;
-use crate::rpc::Request;
 use crate::syncerd::opts::Opts;
 use crate::syncerd::runtime::SyncerdTask;
 use crate::syncerd::runtime::Synclet;
